fixed to for loop
