+++ main.cpp
@@ -1,9 +1,8 @@
 #include "src/server/Server.hpp"
 
-int main()
+int main() 
 {
     Server server(8080);
-
-    server.run();
+    server.start();
     return 0;
 }
